python-sidecar/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn==0.24.0

# PDF processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# OCR and image processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP client
httpx==0.25.2

# Utilities
python-multipart==0.0.6
pydantic==2.5.0

# Vector database and embeddings
pinecone-client==2.2.4
openai==1.3.7
